+++ frontend/app/newsroom.tsx
@@ -7,13 +7,12 @@ import Animated, {
   useSharedValue, 
   useAnimatedStyle, 
   withTiming, 
-  withSpring,
-  interpolate,
-  Extrapolate
+  withSpring
 } from 'react-native-reanimated';
 import Head from 'expo-router/head';
 import { useAuth } from '../util/auth-context';
-import { api, Post } from '../util/api';
+import { api } from '../util/api';
+import { Post } from '../util/types';
 
 export default function Newsroom() {
   const { isAuthenticated, isLoading, user } = useAuth();
